+++ __tests__/demos/3d/index.ts
@@ -4,3 +4,4 @@ export { torus } from './torus';
 export { cylinder } from './cylinder';
 export { force } from './force';
 export { ar } from './webar';
+export { hit_test } from './hit-test';